+++ app.py
@@ -5,13 +5,9 @@ Authored by SC Siao
 """
 
 from flask import *
-from MySQL_con import *
 from flask_cors import CORS
-import jwt
-import time
-# Hash code to save password
-from flask_bcrypt import Bcrypt
-bcrypt = Bcrypt()
+import user_api
+import attractions_api
 
 app=Flask(
 	__name__,
@@ -19,367 +15,9 @@ app=Flask(
     static_url_path="/static"
 )
 CORS(app)
-app.config["JSON_AS_ASCII"]=False
-app.config["TEMPLATES_AUTO_RELOAD"]=True
 
-jwt_key = "key"
-
-@app.route("/api/user/auth", methods=["PUT","GET","DELETE"])
-def auth():
-	if request.method == "GET":
-		try:
-			token = request.cookies.get('token')
-			print("token",token)
-			# Judge with token or not
-			if token == None:
-				time_out_msg = jsonify({
-					"data":None
-				})
-				return time_out_msg,200
-			else:
-				token_data = jwt.decode(token, jwt_key, algorithms="HS256")
-				user_email = token_data["email"]
-				sql_command="""
-				SELECT id,name,email
-				FROM user 
-				WHERE email=%s;
-				"""
-				value_input = (user_email,)
-				user_info = query_data(sql_command,value_input)
-				# Judge token right or wrong
-				if user_info == []:
-					verify_msg = jsonify({
-						"error": True,
-						"message": "token 有誤"
-					})
-				verify_msg = ({
-					"data":{
-						"id":user_info[0][0],
-						"name":user_info[0][1],
-						"email":user_info[0][2]
-					}
-				})
-				print("user_email",verify_msg)
-				return verify_msg,200
-
-		except:
-			errorr_message = {
-				"error": True,
-				"message": "請按照情境提供對應的錯誤訊息"
-			}
-			return errorr_message,500
-	if request.method == "PUT":
-		try:
-			sign_in_data = request.get_json()
-			user_email = sign_in_data["email"]
-			user_password = sign_in_data["password"]
-
-			print("sign_in_data",sign_in_data)
-			# Email check
-			sql_command="""
-			SELECT email, password
-			FROM user 
-			WHERE email=%s;
-			"""
-			value_input = (user_email,)
-			user_check = query_data(sql_command,value_input)
-			email_check = user_check[0][0]
-			passowrd_check = user_check[0][1]
-			
-			# Password with hash coding
-			check_password = bcrypt.check_password_hash(passowrd_check, user_password)
-
-			if check_password == True and len(email_check) == 1:
-				print("go",sign_in_data)
-				result=jsonify({"ok":True})
-				# JWT token sgould not include password
-				sign_in_data_email_only = {
-					"email":sign_in_data["email"]
-				}
-				print("sign_in_data_email_only",sign_in_data_email_only)
-				token = jwt.encode(sign_in_data_email_only, jwt_key, algorithm="HS256")
-				
-				cookie_sustain_days = 1
-				result.set_cookie(key="token", value=token, expires=time.time()+cookie_sustain_days*60*60*24)
-				return result
-			else:
-				error_msg = ""
-				if len(email_check) != 1:
-					error_msg = "信箱"
-				if len(passowrd_check) != 1:
-					error_msg = error_msg+" 密碼"
-				errorr_message = jsonify({
-				"error": True,
-				"message": error_msg+"有誤"
-				})
-				print(errorr_message)
-				return errorr_message,400
-			
-		except:
-			errorr_message = jsonify({
-				"error": True,
-				"message": "請按照情境提供對應的錯誤訊息"
-			})
-			return errorr_message,500
-	if request.method == "DELETE":
-		try:
-			token_del = Response('delete cookies')
-			token_del = jsonify({"ok":True})
-			token_del.set_cookie(key='token', value='', expires=0)
-			
-			print(token_del)
-			return token_del
-		except:
-			errorr_message = {
-				"error": True,
-				"message": "請按照情境提供對應的錯誤訊息"
-			}
-			return errorr_message,500
-@app.route("/api/user", methods=["POST"])
-def user():
-	try:
-		sign_up_data = request.get_json()
-		user_name = sign_up_data["name"]
-		user_email = sign_up_data["email"]
-		user_password = sign_up_data["password"]
-
-		# MySQL data Check
-		# name check
-		sql_command="""
-		SELECT name
-        FROM user 
-		WHERE name=%s;
-		"""
-		value_input = (user_name,)
-		name_check = query_data(sql_command,value_input)
-
-		# Email check
-		sql_command="""
-		SELECT email
-        FROM user 
-		WHERE email=%s;
-		"""
-		value_input = (user_email,)
-		email_check = query_data(sql_command,value_input)
-
-		# Save password with hash coding		
-		user_password = bcrypt.generate_password_hash(password=user_password)
-
-		if name_check == [] and email_check == []:					
-			sql_command = """
-			INSERT INTO user (name, email, password)
-			VALUES (%s,%s,%s);
-			"""
-			value_input = (user_name,user_email,user_password)
-			insert_or_update_data(sql_command,value_input)
-			print("go",sign_up_data)
-			data=jsonify({"ok":True})
-			return data
-
-		else:
-			error_msg = ""
-			if name_check != []:
-				error_msg = "名字"
-			if email_check != []:
-				error_msg = error_msg+" 信箱"
-			errorr_message = jsonify({
-			"error": True,
-			"message": error_msg+"重複"
-			})
-			print(errorr_message)
-			return errorr_message,400
-
-		
-	except:
-		errorr_message = jsonify({
-            "error": True,
-            "message": "請按照情境提供對應的錯誤訊息"
-        })
-		return errorr_message,500
-
-
-@app.route("/api/attractions", methods=["GET"])
-def attractions():
-	# Normal situation
-	try:
-		one_page_quanity=12
-		page=int(request.args.get("page",0))
-		data_start=int(page*one_page_quanity)
-		keyword=request.args.get("keyword",False)
-		if keyword!=False:
-			# attraction information With keyword
-			sql_command="""
-			SELECT * FROM attraction
-			WHERE category like %s OR name like %s
-			LIMIT %s, %s;
-			"""
-			sql_input=(keyword,"%"+keyword+"%",data_start,one_page_quanity)
-			attraction_info=query_data(sql_command,sql_input)
-		else:
-			# attraction information WithOut keyword
-			sql_command="""
-			SELECT * FROM attraction
-			LIMIT %s, %s;
-			"""
-			sql_input=(data_start,one_page_quanity)
-			attraction_info=query_data(sql_command,sql_input)
-		
-		data_result=[]
-		count=0
-		while count<len(attraction_info):
-			data={}
-			data["id"] = attraction_info[count][0]
-			data["name"] = attraction_info[count][1]
-			data["category"] = attraction_info[count][2]
-			data["description"] = attraction_info[count][3]
-			data["address"] = attraction_info[count][4]
-			data["transport"] = attraction_info[count][5]
-			data["mrt"] = attraction_info[count][6]
-			data["lat"] = attraction_info[count][7]
-			data["lng"] = attraction_info[count][8]
-			# attraction image
-			data_image_ls=[]
-			data_image=[]
-			id_check=data["id"]
-			
-			# With keyword 
-			if keyword!=False:
-				sql_command="""
-				SELECT attraction_id,images FROM attraction
-				INNER JOIN attraction_image ON attraction.id=attraction_image.attraction_id
-				WHERE category LIKE %s AND attraction_id LIKE %s OR name LIKE %s AND attraction_id LIKE %s;
-				"""	
-				sql_input=(keyword,id_check,"%"+keyword+"%",id_check)
-				attraction_img=query_data(sql_command,sql_input)
-			# WithOut keyword
-			else:
-				sql_command="""
-				SELECT attraction_id,images FROM attraction
-				INNER JOIN attraction_image ON attraction.id=attraction_image.attraction_id
-				WHERE attraction_id = %s;
-				"""
-				sql_input=(id_check,)
-				attraction_img=query_data(sql_command,sql_input)
-			
-			count_image=0
-			while count_image<len(attraction_img):
-				data_image_ls=[attraction_img[count_image][1]]
-				data_image.extend(data_image_ls)
-				count_image+=1
-			data["images"]=data_image
-			# Merge all data
-			data_result.append(data)
-			# next one
-			count+=1
-	
-		# page judge
-		if len(attraction_info)==one_page_quanity:
-			next_page=page+1
-		else:
-			next_page=None
-		
-		# Combine
-		data_main=jsonify({
-			"nextPage":next_page,
-			"data":data_result
-			})
-		return data_main
-
-	# Unexcept situation
-	except:
-		errorr_message = jsonify({
-            "error": True,
-            "message": "請按照情境提供對應的錯誤訊息"
-        })
-		return errorr_message,500
-
-@app.route("/api/attraction/<attractionId>", methods=["GET"])
-def attractionId(attractionId):
-	# Normal situation
-	try:
-		# attraction information
-		sql_command=f"""
-		SELECT * FROM attraction
-		WHERE id={attractionId};
-		"""
-		attraction_info=query_data_read(sql_command)
-		# With data situation
-		if attraction_info!=[]:
-			data_result=[]
-			count=0
-			data={}
-			data["id"] = attraction_info[count][0]
-			data["name"] = attraction_info[count][1]
-			data["category"] = attraction_info[count][2]
-			data["description"] = attraction_info[count][3]
-			data["address"] = attraction_info[count][4]
-			data["transport"] = attraction_info[count][5]
-			data["mrt"] = attraction_info[count][6]
-			data["lat"] = attraction_info[count][7]
-			data["lng"] = attraction_info[count][8]
-			# attraction image
-			data_image_ls=[]
-			data_image=[]
-			sql_command="""
-			SELECT attraction_id,images FROM attraction
-			INNER JOIN attraction_image ON attraction.id=attraction_image.attraction_id
-			WHERE attraction_id=%s;
-			"""
-			sql_input=(attractionId,)
-			attraction_img=query_data(sql_command,sql_input)
-			count_image=0
-			while count_image<len(attraction_img):
-				data_image_ls=[attraction_img[count_image][1]]
-				data_image.extend(data_image_ls)
-				count_image+=1
-			data["images"]=data_image
-			data_result=data
-			
-			data_main={
-				"data":data_result
-				}
-
-			return data_main
-		# No data situation
-		else:
-			errorr_message = jsonify({
-            "error": True,
-            "message": "請按照情境提供對應的錯誤訊息"
-        })
-		return errorr_message,400
-	# Unexcept situation
-	except:
-		errorr_message = jsonify({
-            "error": True,
-            "message": "請按照情境提供對應的錯誤訊息"
-        })
-		return errorr_message,500
-
-@app.route("/api/categories", methods=["GET"])
-def categories():
-	# Normal situation
-	try:
-		sql_command="""
-		SELECT DISTINCT category FROM attraction;
-		"""
-		categories_raw_data=query_data_read(sql_command)
-		data={}
-		data_categories_ls=[]
-		categories_data=[]
-		count_categories=0
-		while count_categories<len(categories_raw_data):
-			data_categories_ls=[categories_raw_data[count_categories][0]]
-			categories_data.extend(data_categories_ls)
-			count_categories+=1
-		data["data"]=categories_data
-		return data
-	# Unexcept situation
-	except:
-		errorr_message = jsonify({
-            "error": True,
-            "message": "請按照情境提供對應的錯誤訊息"
-        })
-		return errorr_message,500
+app.register_blueprint(user_api.user_api)
+app.register_blueprint(attractions_api.attractions_api)
 
 # Pages
 @app.route("/")
